fix(mq): 被唤醒后条件被抢走改为按剩余超时重试；修正无限等待超时的取值
问题1（丢消息/丢接收）：
  任务从等待中醒来时，队列空间/消息可能已被其它任务抢走。原实现此时直接
  return -1：发送方的消息被悄悄丢掉，接收方在队列明明有数据时被判失败。
  现在引入 svcrt_mq_remain_ms()，按“起始节拍 + 超时上限”计算剩余时间，
  重新排队等待，直到成功或超时（超时返回 1，与既有约定一致）。

问题2（无限等待的取值）：
  svcrt_task_block_in_critical 的参数是 uint32，调用方用 (uint32)(-1) 表示
  无限等待，而原实现用 timeout_ms > 0u 判断，负值被当成 0xFFFFFFFF 毫秒，
  SVCRT_MS_TO_TICK 溢出成一个不可控的巨大节拍数。改为先转 int32 再判断。

顺带修正 tools/pack_app.py --load-addr 的帮助文本（默认值已改为负载基址）。

验证：armclang -fsyntax-only 通过；未上板。

+++ kernelsrc/src/svcrt_task.c
@@ -845,7 +845,10 @@ int32 svcrt_task_block_in_critical(uint32 timeout_ms)
 
     p_tsk = &svcrt_task_table[svcrt_current_task_id - 1];
     p_tsk->wake_reason = 0;
-    p_tsk->wait_time   = (timeout_ms > 0u) ? (int32)SVCRT_MS_TO_TICK(timeout_ms) : -1;
+    /* 参数类型是 uint32，调用方用 (uint32)(-1) 表示无限等待；
+     * 必须先转成有符号再判断，否则负值会被当成 0xFFFFFFFF 毫秒，
+     * MS_TO_TICK 溢出成一个不可控的巨大节拍数。 */
+    p_tsk->wait_time   = ((int32)timeout_ms > 0) ? (int32)SVCRT_MS_TO_TICK((uint32)timeout_ms) : -1;
     p_tsk->status      = SVCRT_TASK_WAIT;
 
     SVCRT_SWITCH_TASK();                /* 只是置 PendSV pending，此刻中断还关着 */
